+++ src/components/offer/offer-description/offer-description.tsx
@@ -1,7 +1,7 @@
 import AddToBookmarks from '@/components/add-to-bookmarks-btn/add-to-bookmarks-btn';
 import { OfferFull } from '../../../libs/types/types';
 import HostInfo from '../host-info/host-info';
-import OfferReviews from '../offer-reviews/offer-reviews';
+import OfferReviews from '../reviews/offer-reviews/offer-reviews';
 import { BookmarkOfferFull, OfferType } from '@/libs/const';
 
 type OfferDescription = {
