--- src/components/offer/offer-reviews/offer-reviews.tsx
@@ -1,36 +1,0 @@
-import ReviewForm from '../offer/review-form/review-form';
-
-function OfferReviews() {
-  return (
-    <section className="offer__reviews reviews">
-      <h2 className="reviews__title">Reviews · <span className="reviews__amount">1</span></h2>
-      <ul className="reviews__list">
-        <li className="reviews__item">
-          <div className="reviews__user user">
-            <div className="reviews__avatar-wrapper user__avatar-wrapper">
-              <img className="reviews__avatar user__avatar" src="img/avatar-max.jpg" width={54} height={54} alt="Reviews avatar" />
-            </div>
-            <span className="reviews__user-name">
-              Max
-            </span>
-          </div>
-          <div className="reviews__info">
-            <div className="reviews__rating rating">
-              <div className="reviews__stars rating__stars">
-                <span style={{width: '80%'}} />
-                <span className="visually-hidden">Rating</span>
-              </div>
-            </div>
-            <p className="reviews__text">
-              A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam. The building is green and from 18th century.
-            </p>
-            <time className="reviews__time" dateTime="2019-04-24">April 2019</time>
-          </div>
-        </li>
-      </ul>
-      <ReviewForm/>
-    </section>
-  );
-}
-
-export default OfferReviews;
