+++ src/components/offer/reviews/reviews-list/reviews-list.tsx
@@ -1,0 +1,11 @@
+import ReviewItem from '../review-item/review-item';
+
+function ReviewsList() {
+  return (
+    <ul className="reviews__list">
+      <ReviewItem/>
+    </ul>
+  );
+}
+
+export default ReviewsList;
